Move cache tests.
