+++ fxa/tests/test_oauth.py
@@ -5,11 +5,11 @@
 import json
 import mock
 import responses
-import time
 import six
 
 import fxa.errors
-from fxa.oauth import Client, scope_matches, MemoryCache
+from fxa.cache import MemoryCache
+from fxa.oauth import Client, scope_matches
 
 from fxa.tests.utils import unittest
 
@@ -342,21 +342,6 @@ class TestScopeMatch(unittest.TestCase):
         self.assertFalse(scope_matches(['abc:xyz'], 'abc'))
         self.assertFalse(scope_matches(['abc:xyz'], ['abc']))
         self.assertFalse(scope_matches(['abc:xyz', 'def'], ['abc', 'def']))
-
-
-class TestMemoryCache(unittest.TestCase):
-    def setUp(self):
-        self.cache = MemoryCache()
-
-    def test_can_get_what_has_been_set(self):
-        self.cache.set('Foo', 'Bar')
-        self.assertEqual(self.cache.get('Foo'), 'Bar')
-
-    def test_expires(self):
-        self.cache = MemoryCache(0.01)
-        self.cache.set('Foo', 'Bar')
-        time.sleep(0.01)
-        self.assertIsNone(self.cache.get('Foo'))
 
 
 class TestCachedClient(unittest.TestCase):
